+++ src/lib.rs
@@ -1,13 +1,13 @@
 pub mod builder;
 pub mod cli;
 pub mod client;
+pub mod commons;
 pub mod crd;
 pub mod error;
 pub mod label_selector;
 pub mod labels;
 pub mod logging;
 pub mod namespace;
-pub mod opa;
 pub mod pod_utils;
 pub mod product_config_utils;
 pub mod resources;